requirements.txt
javalang==0.13.0
langchain-core
langchain>=0.1.0
langchain-openai>=0.1.0
python-dotenv>=1.0.0
openai>=1.3.0
javalang>=0.13.0
xmltodict>=0.13.0
pytest>=7.4.0
tiktoken